Add some default implementation methods
